+++ src/tagoapi/__init__.py
@@ -1,5 +1,6 @@
 from ._version import __version__
 from .client import TAGOClient
+from .station_catalog import StationCatalog
 from .models import ArrivalInfo, BaseModel, CityCode, Route, Station, Vehicle
 from .resources import (
     ArrivalResource,
@@ -12,6 +13,7 @@ from .resources import (
 
 __all__ = [
     "TAGOClient",
+    "StationCatalog",
     "__version__",
     "BaseModel",
     "Route",