+++ src/tagoapi/utils/__init__.py
@@ -1,4 +1,0 @@
-"""Local data utilities.
-
-정류소 CSV 검색은 추후 StationCatalog로 제공한다.
-"""